requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.3.5
python-dotenv==1.0.0
qdrant-client==1.7.0
pymysql==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
apscheduler==3.10.4
pydantic==2.4.2
numpy==1.26.1
cryptography==41.0.5
lxml==4.9.3
asyncio==3.4.3
httpx==0.25.1
python-multipart==0.0.6
aiohttp==3.9.1
loguru==0.7.2
tabulate==0.9.0
fuzzywuzzy==0.18.0
SQLAlchemy==2.0.36
alembic==1.14.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
psycopg2-binary==2.9.10
pytest==8.3.3
email-validator==2.2.0
geoip2==4.8.0
pycountry==24.6.1